comments and unified test_init starting sequence
